requirements.txt
# Core RAG Components
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
langchain>=0.1.0
langchain-community>=0.0.10

# Document Processing - Python 3.13 Compatible
PyPDF2>=3.0.0
pdfplumber>=0.9.0
Markdown>=3.5.0
markdownify>=0.11.0

# CSV and Data Processing
pandas>=2.0.0

# Image Processing (Optional - skip for now if issues)
# pytesseract>=0.3.10
# Pillow>=10.0.0
# opencv-python>=4.8.0

# LLM Integration (Optional - install later)
# openai>=1.3.0
# anthropic>=0.7.0

# Essential Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
PyYAML>=6.0.0
tqdm>=4.66.0

# Development and Testing
jupyter>=1.0.0

# Web Interface (Optional - install later)
# streamlit>=1.28.0
# gradio>=4.0.0

# Additional Dependencies for Content Processing
beautifulsoup4>=4.12.0
